Updated default constructor for `Url`
By default it no longer takes optional parameters for creating SEO-friendly URLs (defaults to making them so), and will now default to making the URL the canonical one. The reasoning behind this is to remove the single responsibility principle violation. Forcing the code to instead use functions for overriding these also makes the code more readable.